# docs/swagger.yaml
basePath: /api
definitions:
  github_com_amalgamated-tools_biblioteka_internal_db.Timestamp:
    properties:
      time.Time:
        type: string
    type: object
...
  internal_handlers.authorDTO:
    properties:
      created_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
      goodreads_id:
        type: string
      google_books_id:
        type: string
      hardcover_id:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      updated_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
    type: object
...
  internal_handlers.bookDTO:
    properties:
      asin:
        type: string
      authors:
        items:
          $ref: '#/definitions/internal_handlers.authorDTO'
        type: array
      cover_image_url:
        type: string
      created_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
      description:
        type: string
      files:
        items:
          $ref: '#/definitions/internal_handlers.bookFileDTO'
        type: array
      goodreads_id:
        type: string
      google_books_id:
        type: string
      hardcover_id:
        type: string
      id:
        type: string
      isbn10:
        type: string
      isbn13:
        type: string
      language:
        type: string
      num_pages:
        type: integer
      publication_date:
        type: string
      publisher:
        type: string
      series:
        items:
          $ref: '#/definitions/internal_handlers.bookSeriesEntryDTO'
        type: array
      title:
        type: string
      updated_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
    type: object
  internal_handlers.bookFileDTO:
    properties:
      book_id:
        type: string
      created_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
      file_hash:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      id:
        type: string
      updated_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
    type: object
  internal_handlers.bookRequest:
    properties:
      asin:
        type: string
      cover_image_url:
        type: string
      description:
        type: string
      goodreads_id:
        type: string
      google_books_id:
        type: string
      hardcover_id:
        type: string
      isbn10:
        type: string
      isbn13:
        type: string
      language:
        type: string
      num_pages:
        type: integer
      publication_date:
        type: string
      publisher:
        type: string
      title:
        type: string
    type: object
  internal_handlers.bookSeriesEntryDTO:
    properties:
      position:
        type: number
      series:
        $ref: '#/definitions/internal_handlers.seriesDTO'
    type: object
  internal_handlers.bookSummaryDTO:
    properties:
      asin:
        type: string
      cover_image_url:
        type: string
      created_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
      description:
        type: string
      goodreads_id:
        type: string
      google_books_id:
        type: string
      hardcover_id:
        type: string
      id:
        type: string
      isbn10:
        type: string
      isbn13:
        type: string
      language:
        type: string
      num_pages:
        type: integer
      publication_date:
        type: string
      publisher:
        type: string
      title:
        type: string
      updated_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
    type: object
...
  internal_handlers.errorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  internal_handlers.seriesDTO:
    properties:
      created_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
      goodreads_id:
        type: string
      google_books_id:
        type: string
      hardcover_id:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Personal library management API
  title: Biblioteka API
  version: "1.0"
paths:
...
  /books:
    get:
      description: Returns all books (summary without relations)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.bookSummaryDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: List books
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: Create a new book
      parameters:
      - description: Book data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.bookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.bookDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a book
      tags:
      - Books
  /books/{id}:
    delete:
      description: Delete a book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a book
      tags:
      - Books
    get:
      description: Returns a single book with authors, series, and files
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.bookDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a book
      tags:
      - Books
    put:
      consumes:
      - application/json
      description: Update an existing book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Book data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.bookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.bookDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a book
      tags:
      - Books
  /books/{id}/authors:
    get:
      consumes:
      - application/json
      description: Get the list of authors for a book, or replace them
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.authorDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get or set book authors
      tags:
      - Books
    put:
      consumes:
      - application/json
      description: Get the list of authors for a book, or replace them
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.authorDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get or set book authors
      tags:
      - Books
  /books/{id}/files:
    get:
      consumes:
      - application/json
      description: List files for a book, or add a new file
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.bookFileDTO'
            type: array
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.bookFileDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get or add book files
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: List files for a book, or add a new file
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.bookFileDTO'
            type: array
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.bookFileDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get or add book files
      tags:
      - Books
  /books/{id}/series:
    get:
      consumes:
      - application/json
      description: Get the list of series for a book, or replace them
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.bookSeriesEntryDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get or set book series
      tags:
      - Books
    put:
      consumes:
      - application/json
      description: Get the list of series for a book, or replace them
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.bookSeriesEntryDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get or set book series
      tags:
      - Books
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"